requirements.txt
alembic==1.4.0
google-auth==1.11.0
google-cloud-error-reporting==0.33.0
google-cloud-logging==1.14.0
Jinja2==2.11.1
numpy==1.18.1
Orange3==3.24.1
pandas==1.0.1
psycopg2-binary==2.8.4
pyfakefs==3.7.1
pytest==5.3.5
PyYAML==5.3
scikit-posthocs==0.6.2
scipy==1.4.1
seaborn==0.10.0
sqlalchemy==1.3.13

# Needed for development.
pylint==2.4.4
pytype==2019.11.27
yapf==0.28.0
